How to save aliases when using fish shell
First, apologies if I have submitted two pull requests, I'm not a developer and this is my first time contributing to something that isn't my personal repository.

This is something I initially found confusing. I installed fish shell based on a recommendation online and didn't realize that the alias command behaves differently. It creates the alias just as bash would but only for the current session. I think that many newcomers that happen to have fish installed may become confused that when their aliases are gone the next session.

If you are interested it may be nice to add a section on fish shell here as I personally find it a great "quality of life tweak". If you are interested let me know and I can draft something up.